Object Pooling and Chunks
Object Pools have been built and can be created within the editor. Use
of pooling at runtime not yet supported.
Chunks are now able to be constructed within the editor also. A Chunk
defines a layout for a set of objects that can be loaded into or from
the editor.
Note: Objects cannot be a part of a Chunk if they are not listed in the
ObjectPooler's Object Pools list.